add clean cid interface

+++ src/coroutine.h
@@ -1,7 +1,7 @@
 #ifndef __COROUTINE_H__
 #define __COROUTINE_H__
 
-typedef int coroutine_t;
+typedef unsigned long coroutine_t;
 
 int coroutine_create(coroutine_t *cidp, const void *attr,
     void*(*start_rtn)(void*), void *arg);
